Safari-611.1.10/LayoutTests/fast/css3-text/css3-text-decoration/text-decoration-thickness-parse-expected.txt

PASS window.getComputedStyle(document.getElementById('test1')).getPropertyValue('text-decoration-thickness') is "auto"
PASS window.getComputedStyle(document.getElementById('test2')).getPropertyValue('text-decoration-thickness') is "auto"
PASS window.getComputedStyle(document.getElementById('test3')).getPropertyValue('text-decoration-thickness') is "from-font"
PASS window.getComputedStyle(document.getElementById('test4')).getPropertyValue('text-decoration-thickness') is "auto"
PASS window.getComputedStyle(document.getElementById('test5')).getPropertyValue('text-decoration-thickness') is "auto"
PASS window.getComputedStyle(document.getElementById('test6')).getPropertyValue('text-decoration-thickness') is "37px"
PASS window.getComputedStyle(document.getElementById('test7')).getPropertyValue('text-decoration-thickness') is "200px"
PASS window.getComputedStyle(document.getElementById('test8')).getPropertyValue('text-decoration-thickness') is "40px"
PASS document.getElementById('target').sheet.cssRules[0].style.textDecorationThickness is ""
PASS document.getElementById('target').sheet.cssRules[1].style.textDecorationThickness is "auto"
PASS document.getElementById('target').sheet.cssRules[2].style.textDecorationThickness is "from-font"
PASS document.getElementById('target').sheet.cssRules[3].style.textDecorationThickness is ""
PASS document.getElementById('target').sheet.cssRules[4].style.textDecorationThickness is ""
PASS document.getElementById('target').sheet.cssRules[5].style.textDecorationThickness is "37px"
PASS document.getElementById('target').sheet.cssRules[6].style.textDecorationThickness is "2em"
PASS successfullyParsed is true

TEST COMPLETE
